- moved explore screen logic to pages directory - improved url state for explore screen
